fix: map select now not jitter rendering

+++ src/ui/DetailDisplay.cpp
@@ -93,7 +93,7 @@ SDL_FPoint DetailDisplay::getPosition()
 void DetailDisplay::render() 
 {
     // draws border
-    SDL_FRect borderRect = { mPosition.x, mPosition.y, width, height };
+    SDL_FRect borderRect = { mPosition.x, mPosition.y, static_cast<float>(width), static_cast<float>(height) };
     SDL_SetRenderDrawColor(gRenderer, 0xAA, 0xAA, 0xAA, 0xFF);
     SDL_RenderFillRect(gRenderer, &borderRect);
 
@@ -101,8 +101,8 @@ void DetailDisplay::render()
     SDL_FRect innerRect = {
         mPosition.x + DETAIL_BORDER_WIDTH,
         mPosition.y + DETAIL_BORDER_WIDTH,
-        width - (2 * DETAIL_BORDER_WIDTH),
-        height - (2 * DETAIL_BORDER_WIDTH)
+        static_cast<float>(width - (2 * DETAIL_BORDER_WIDTH)),
+        static_cast<float>(height - (2 * DETAIL_BORDER_WIDTH))
     };
     SDL_SetRenderDrawColor(gRenderer, 0xCC, 0xCC, 0xCC, 0xFF);
     SDL_RenderFillRect(gRenderer, &innerRect);
