+++ src/ui/DetailDisplayComponent.cpp
@@ -4,7 +4,6 @@
  * @author Denmar Ermitano
  * @date 2025-03-02
  */
-
 #include <ui/DetailDisplayComponent.h>
 #include <Global.h>
 
@@ -52,7 +51,7 @@ void DetailDisplayComponent::setComponentPosition(float x, float y)
 void DetailDisplayComponent::render() 
 {
     // draws rectangle for border
-    SDL_FRect borderRect = { xPosition, yPosition, width, DETAIL_COMPONENT_HEIGHT };
+    SDL_FRect borderRect = { xPosition, yPosition, static_cast<float>(width), static_cast<float>(DETAIL_COMPONENT_HEIGHT) };
     SDL_SetRenderDrawColor(gRenderer, 0x77, 0x77, 0x77, 0xFF);
     SDL_RenderFillRect(gRenderer, &borderRect);
 
@@ -61,7 +60,7 @@ void DetailDisplayComponent::render()
         xPosition + COMPONENT_BORDER_WIDTH,
         yPosition + COMPONENT_BORDER_WIDTH,
         width - (2 * COMPONENT_BORDER_WIDTH),
-        DETAIL_COMPONENT_HEIGHT - (2 * COMPONENT_BORDER_WIDTH)
+        static_cast<float>(DETAIL_COMPONENT_HEIGHT - (2 * COMPONENT_BORDER_WIDTH))
     };
     SDL_SetRenderDrawColor(gRenderer, 0xFF, 0xFF, 0xFF, 0xFF);
     SDL_RenderFillRect(gRenderer, &innerRect);
